requirements.txt
fastapi
pytest
uvicorn
httpx
supabase
python-dotenv

